update delay moved into service
