+++ app/models/device.rb
@@ -1,5 +1,5 @@
 class Device < ApplicationRecord
-  include SlackNotifiable, HumanizeMinutes, RemoveId
+  include SlackNotifiable, RemoveId
 
   belongs_to :user
   has_one :config, dependent: :destroy
@@ -93,21 +93,9 @@ class Device < ApplicationRecord
     "A new device was created, id: #{id}, name: #{name}, user_id: #{user_id}. There are now #{Device.count} devices"
   end
 
-  def update_delay(mins)
-    mins.to_i.zero? ? update(delayed: nil) : update(delayed: mins)
-  end
-
   def switch_fog
     update(fogged: !fogged)
     fogged
-  end
-
-  def humanize_delay
-    if delayed.nil?
-      "#{name} is not delayed."
-    else
-      "#{name} delayed by #{humanize_minutes(delayed)}."
-    end
   end
 
   def public_info
