+++ app/services/users/devices/update_device.rb
@@ -9,7 +9,7 @@ module Users::Devices
 
     def update_device
       if @params[:delayed]
-        @device.update_delay(@params[:delayed])
+        update_delay(@params[:delayed])
       elsif @params[:published]
         @device.update(published: !@device.published)
       elsif @params[:name]
@@ -42,6 +42,31 @@ module Users::Devices
 
     def boolean_to_state(boolean)
       boolean ? 'on' : 'off'
+    end
+
+    def update_delay(mins)
+      mins.to_i.zero? ? @device.update(delayed: nil) : @device.update(delayed: mins)
+    end
+
+    def humanize_delay
+      if @device.delayed.nil?
+        "#{@device.name} is not delayed."
+      else
+        "#{@device.name} delayed by #{humanize_minutes(@device.delayed)}."
+      end
+    end
+
+    def humanize_minutes(minutes)
+      if minutes < 60
+        "#{minutes} #{'minute'.pluralize(minutes)}."
+      elsif minutes < 1440
+        hours = minutes / 60
+        minutes = minutes % 60
+        "#{hours} #{'hour'.pluralize(hours)} and #{minutes} #{'minutes'.pluralize(minutes)}."
+      else
+        days = minutes / 1440
+        "#{days} #{'day'.pluralize(days)}."
+      end
     end
   end
 end
